+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Roughness and metalness maps not being assigned correctly.
 - Case where textures using shared "Source" with different encodings were not treated as unique.
 
+### Changed
+- Move random functions around.
+
 ## [0.0.8] - 2022-12-11
 ### Fixed
 - Three.js semver package version.